\name{affycoretools-defunct}
\alias{affycoretools-defunct}
\title{Defunct functions in package \pkg{affycoretools}}
\alias{annBM}
\alias{vennSelectBM}
\alias{limma2biomaRt}
\alias{probes2tableBM}
\alias{foldFiltBM}
\alias{foldFilt}
\alias{vennSelectFC}
\alias{hyperG2annaffy}
\alias{hyperGoutput}
\alias{limma2annaffy}
\alias{probes2table}
\alias{vennSelect}
\description{Functions that have been removed from affycoretools, as
  they are no longer needed.}
\usage{
#These functions are defunct as of affycoretools 1.40.0
annBM
vennSelectBM
limma2biomaRt
probes2tableBM
foldFiltBM
foldFilt
vennSelectFC
hyperG2annaffy
hyperGoutput
limma2annaffy
probes2table
vennSelect}
\details{
Most of these functions have been superceded by functions intended to
use ReportingTools instead of annaffy. Please see the Refactoring
Affycoretools vignette for more information.
}
\keyword{internal}