% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dataAndHeatmapPage}
\alias{dataAndHeatmapPage}
\title{A function to create an annotated HTML table for all genes in a significant
gene set as well as a heatmap of these data.}
\usage{
dataAndHeatmapPage(eset, fit, ind, columns = NULL, annot = NULL, fname,
  heatmap, title, key = TRUE, fitind = NULL)
}
\arguments{
\item{eset}{An \code{\link{ExpressionSet}} containing normalized, summarized
gene expression data.}

\item{fit}{An \code{MArrayLM} object.}

\item{ind}{Numeric vector indicating which rows of the
\code{\link{ExpressionSet}} to use.}

\item{columns}{Numeric vector indicating which columns of the
\code{\link{ExpressionSet}} to use. If \code{NULL}, all columns will be
used.}

\item{annot}{Character. The name of the annotation package for the array
used. If \code{NULL}, data will be output without any additional annotation.}

\item{fname}{The filename of the resulting output, without the 'html' file
extension.}

\item{heatmap}{Character. The filename of the heatmap to append to the
bottom of the HTML page.}

\item{title}{Title to be placed at the top of the resulting HTML page.}

\item{key}{Character. The filename of the heatmap key to append to the
bottom of the HTML page.}

\item{fitind}{Numeric. Which column of the \code{MArrayLM} object to use for
output in the HTML table.}
}
\description{
This is intended to be an internal function to \code{runRomer}. It is
documented here only because it may be necessary to pass alternative
arguments to this function from \code{runRomer}.
}
\details{
This function creates an annotation table using \code{probes2table} if an
annotation file is used, otherwise data will be output in a simple HTML
table. A heatmap showing the expression values for all the genes in the gene
set is then placed below this table, along with a key that indicates the
range of the expression values.
}
\author{
James W. MacDonald
}

