% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractStats}
\alias{extractStats}
\title{Extract Statistics from an MArrayLM object}
\usage{
extractStats(prbs, fit, comp, statistics)
}
\arguments{
\item{prbs}{A vector of probe IDs}

\item{fit}{An \code{\link[limma:marraylm]{MArrayLM}} object}

\item{comp}{Numeric vector of length one, designating which comparison to
select from the \code{\link[limma:marraylm]{MArrayLM}} object}

\item{statistics}{Any or all of 'tstat', 'pval', 'FC', used to control the
output of statistics from the \code{\link[limma:marraylm]{MArrayLM}} object}
}
\value{
A named list that can have up to three vectors of statistics.
}
\description{
This is a utility function that can be used to extract the statistics for a
set of probesets from an \code{\link[limma:marraylm]{MArrayLM}} object. This
may be useful for others, but currently is considered an internal function
and is not intended to be called by end users.
}
\details{
This is an internal function used to return a named list of statistics,
which can then be passed to \code{probes2table} for creating HTML and/or
text tables.
}
\author{
James W. MacDonald <jmacdon@u.washington.edu>
}
\keyword{internal}

