% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getIndex}
\alias{getIndex}
\title{Get Indices}
\usage{
getIndex(comps, groups)
}
\arguments{
\item{comps}{A list of comparisons}

\item{groups}{A vector of groupnames}
}
\description{
This is an internal function and not intended to be called by end users.
}
\author{
James W. MacDonald <jmacdon@u.washington.edu>
}
\keyword{internal}

