% Generated by roxygen2 (4.0.1): do not edit by hand
\name{houtSel}
\alias{houtAll}
\alias{houtSel}
\alias{houtSplit}
\title{Internal functions for hyperGoutput}
\usage{
houtSel(tab, eset, fit, comp, statistics, html, text)
}
\arguments{
\item{tab}{A \code{data.frame}, resulting from a call to
\code{\link[GOstats]{probeSetSummary}}}

\item{eset}{An \code{link[Biobase:class.ExpressionSet]{ExpressionSet}}
object}

\item{fit}{An \code{\link[limma:marraylm]{MArrayLM}} object, resulting from
a call to \code{\link[limma:ebayes]{eBayes}}}

\item{comp}{Numeric vector of length one, designating which comparison to
select from the \code{\link[limma:marraylm]{MArrayLM}} object}

\item{statistics}{Any or all of 'tstat', 'pval', 'FC', used to control the
output of statistics from the \code{\link[limma:marraylm]{MArrayLM}} object}

\item{html}{Boolean. Output HTML tables? Defaults to \code{TRUE}}

\item{text}{Boolean. Output text tables? Defaults to \code{TRUE}}
}
\value{
Nothing is returned. These functions are called solely for their
side effect, the output of HTML and/or text tables.
}
\description{
These functions are internal functions for \code{hyperGoutput} and are not
intended to be called directly by the end user.
}
\details{
These functions control the output and ordering of the probesets for the
\code{hyperGoutput} function.
}
\author{
James W. MacDonald <jmacdon@u.washington.edu>
}
\keyword{internal}

