% Generated by roxygen2 (4.0.1): do not edit by hand
\name{annBM}
\alias{annBM}
\alias{linksBM}
\title{Select Available Annotation from a Biomart}
\usage{
annBM(mart, annot, species)
}
\arguments{
\item{mart}{A 'mart' connection, typically created by a call to
\code{\link[biomaRt]{useMart}}.}

\item{annot}{A character vector of annotation sources. This is not typically
useful for an end user to specify.}

\item{species}{A species name, of the form e.g., 'hsapiens'}
}
\value{
Normally called by an end user to output a character vector of
annotation sources.
}
\description{
These functions are designed to do two things that are useful for an end
user. If called with no arguments, they will output a character vector of
annotation sources that are typically available from a Biomart database. If
called with a 'mart' connection (typically created by a call to
\code{\link[biomaRt]{useMart}}), they will return a character vector of
annotation sources that exist for that particular Biomart and species. If
called with a 'mart' connection and a character vector of annotation
sources, they will return a list that is intended to be used by other
functions for creating HTML pages. This last function doesn't have any real
utility for the end user.
}
\details{
The purpose of these functions is to either give an example of typical
annotation sources that may be available at a particular Biomart, or to
output those sources that are known to exist at a Biomart.

\code{linksBM} is intended to list those annotation sources that may be
turned into hyperlinks whereas \code{annBM} is intended to list those
annotation sources that will not be linked.

These functions have only a few of the possible annotation sources, and
currently there is no simple way to extend these sources. Additions to the
list are possible, however. Please contact me if there is something in
particular that should be included in either list.
}
\examples{
annBM()
}
\author{
James W. MacDonald <jmacdon@u.washington.edu>
}
\keyword{manip}

