% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vennSelect.R
\name{getCols}
\alias{getCols}
\title{Correct Ordering of Contrasts}
\usage{
getCols(design, contrast)
}
\arguments{
\item{design}{A design matrix, usually produced by a call to
\code{model.matrix}.}

\item{contrast}{A contrasts matrix, either produced by hand or as a result
of a call to \code{\link[limma]{makeContrasts}}}
}
\value{
This function returns the correct order for the contrasts.
}
\description{
A function to determine the correct ordering of contrasts for
\code{vennSelect}. This is an internal function and should not be called by
the end user
}
\author{
James W. MacDonald
}
\keyword{internal}

