\name{getIndex}
\alias{getIndex}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Get Indices }
\description{
  This is an internal function and not intended to be called by end users.
}
\usage{
getIndex(comps, groups)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{comps}{A list of comparisons }
  \item{groups}{ A vector of groupnames}
}


\author{James W. MacDonald <jmacdon@med.umich.edu> }
\keyword{ internal }% __ONLY ONE__ keyword per line
