% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennSelect.R
\name{intNames}
\alias{intNames}
\title{Create Names for Venn Diagram Intersections}
\usage{
intNames(x)
}
\arguments{
\item{x}{A \code{TestResults} object from a call to \code{decideTests}.}
}
\value{
A vector of names.
}
\description{
This function is designed to create the names for all the intersections of a
Venn diagram based on the names of all the single comparisons. This is an
internal function and is not intended to be called by the end user.
}
\author{
James W. MacDonald <jmacdon@u.washington.edu>
}
\keyword{internal}
