% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affystart.R
\name{make.cl}
\alias{make.cl}
\title{A Function to Make a Classlabel Vector for Plotting}
\usage{
make.cl(filenames)
}
\arguments{
\item{filenames}{A vector of .cel filenames}
}
\value{
A vector of numbers to be used for plotting colors and line
types
}
\description{
This function takes a vector of filenames and makes a classlabel vector for
plotting functions of \code{affystart}. This is an internal function and is
not intended to be called by the end user.
}
\author{
James W. MacDonald <jmacdon@u.washington.edu>
}
\keyword{internal}
