% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vennSelect.R
\name{makeIndices}
\alias{makeIndices}
\title{Create Indices for Venn Diagrams}
\usage{
makeIndices(x, method = "same")
}
\arguments{
\item{x}{A \code{TestResults} object from a call to \code{decideTests}.}

\item{method}{One of "same", "up", "down", "both", indicating direction of
differential expression. See details of \code{vennSelect} for more
information.}
}
\value{
A list containing \code{TRUE} and \code{FALSE} values, to be used by
\code{vennSelect}.
}
\description{
This function is used to create indices for making Venn diagrams or
\code{vennCounts} objects. This is an internal method and is not intended to
be called by the end user.
}
\author{
James W. MacDonald <jmacdon@u.washington.edu>
}
\keyword{internal}

