% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotma.R
\name{maplot}
\alias{maplot}
\title{A Function to make MA plots from all arrays.}
\usage{
maplot(object)
}
\arguments{
\item{object}{An ExpressionSet or matrix containing log-transformed array
data.}
}
\value{
No output. Used only for the side effect of creating MA plots.
}
\description{
This function creates an MA plot for all arrays in either an ExpressionSet
or a matrix. A 'baseline' array is created using the median expression for
each gene, and each array is then compared to the baseline array.
}
\author{
James W. MacDonald <jmacdon@u.washington.edu>
}
\keyword{hplot}
